Move fetch_departm in another module, calculate dept number automatically
